Drop the project row from the consent screen
Naming one project made sense while a connection was confined to it. Now that it can
act wherever its user can, the row states a fact and the bullet beneath immediately
says that fact is not a limit — and a reader scanning a consent screen will take a named
project as a limit. Removing it leaves one honest line about projects instead of two
that argue.

With nothing displaying them, the projectId and projectName fields on the consent
response and the describeTargetProject helper behind them are dead, so they go too.
Redemption already refuses with a precise reason when a user has no reachable project,
which was the only enforcement that helper contributed. consent-details.ts is left
holding just the project listing, so it is renamed for what it does.
