+++ packages/server/api/cache/codes/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "codes",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "commonjs",
+  "dependencies": {
+    "@tsconfig/node20": "20.1.4",
+    "@types/node": "20.14.8",
+    "typescript": "5.6.3"
+  }
+}
