--- packages/server/api/src/app/oauth/consent-details.ts
@@ -1,76 +1,0 @@
-import { isNil } from '@openops/shared';
-import { projectService } from '../project/project-service';
-import { userService } from '../user/user-service';
-import { getOAuthProjectMembershipService } from './project-membership-factory';
-
-export type ConsentProject = {
-  projectId: string;
-  projectName: string;
-};
-
-/**
- * Every project the user may act in, for a client deciding where to switch to.
- *
- * Names are resolved here rather than by the membership service because the seam
- * answers questions about authority, not about display.
- */
-export async function listAvailableProjects(
-  userId: string,
-): Promise<ConsentProject[]> {
-  const user = await userService.get({ id: userId });
-
-  if (isNil(user)) {
-    return [];
-  }
-
-  const memberships = await getOAuthProjectMembershipService().listForUser(
-    user,
-  );
-  const projects = await Promise.all(
-    memberships.map((membership) =>
-      projectService.getOne(membership.projectId),
-    ),
-  );
-
-  return memberships.map((membership, index) => ({
-    projectId: membership.projectId,
-    projectName: projects[index]?.displayName ?? membership.projectId,
-  }));
-}
-
-/**
- * The project a new connection would be bound to, resolved for display only.
- *
- * The binding itself happens when the authorization code is redeemed, from the same
- * membership lookup. Naming the project on the consent screen is what lets the user see
- * whose data they are about to hand over.
- *
- * Absence is not an error here. Redemption performs the same lookup and refuses with a
- * precise reason, which serves the user better than a half-rendered consent screen.
- */
-export async function describeTargetProject(
-  userId: string,
-): Promise<ConsentProject | null> {
-  const user = await userService.get({ id: userId });
-
-  if (isNil(user)) {
-    return null;
-  }
-
-  const membership = await getOAuthProjectMembershipService().getDefaultForUser(
-    user,
-  );
-
-  if (isNil(membership)) {
-    return null;
-  }
-
-  const project = await projectService.getOne(membership.projectId);
-
-  return {
-    projectId: membership.projectId,
-    // Falling back to the id keeps the screen honest if the project is unreadable:
-    // it still names what is being granted rather than showing nothing.
-    projectName: project?.displayName ?? membership.projectId,
-  };
-}
